put InferWidths in a seperate file and fix spaces
